Rename file and restore some things
